Allow images to be removed
Images need a onClickListener with an overlay to show what a second press will do. README updated to show what the goals are. A redesign is required for editing submissions and showing a activity when you have already submitted and what you need to know (e.g. voting info and times)